+++ activerecord/CHANGELOG.md
@@ -26,5 +26,50 @@
 
     *Michael Duchemin*
 
+*   Add more helpful methods for ActiveRecord::Enum.
+
+    *Jason Lee*
+
+    ```rb
+    class Book
+      enum status: %i[draft published archived]
+    end
+
+    Book.status_options # => [["Drafting", "draft"], ["Published", "published"], ["Archived", "archived"]]
+
+    @book = Book.new(status: :draft)
+    @book.status # => "draft"
+    @book.status_name # => "Drafting"
+    @book.status_color # => "#999999"
+    @book.status_value # => 0
+
+    @book.status = :published
+    @book.status_name # => "Published"
+    @book.status_color # => "green"
+    @book.status_value # => 1
+    ```
+
+    Custom name, color in I18n config:
+
+    ```yml
+    en:
+      activerecord:
+        enums:
+          book:
+            status:
+              draft: Drafting
+              published: Published
+              archived: Archived
+            status_color:
+              draft: "#999999"
+              published: "green"
+              archived: "red"
+    ```
+
+    `status_options` for select tag in Views:
+
+    ```erb
+    <%= f.select :status, Book.status_options %>
+    ```
 
 Please check [6-0-stable](https://github.com/rails/rails/blob/6-0-stable/activerecord/CHANGELOG.md) for previous changes.